ANDROID: Remove ABI Monitoring on android-mainline
We stabilized the ABI tooling on the GKI branches by now. There is no
need anymore to keep the ABI Monitoring alive on this branch. It rather
causes friction as the ABI changes all the time and the purpose of
battletesting the tooling is not as valid anymore as before.
(Also, the representation diffs slowly blow up the repository.)

Bug: 111449103
Change-Id: I86f0baf48f2f183f25aa63e1f560e8b767e7e03f
Signed-off-by: Matthias Maennich <maennich@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>